[WPE][GTK] visible rect in case animations is not correctly calculated https://bugs.webkit.org/show_bug.cgi?id=280612
Reviewed by Miguel Gomez.

For animations starting partially outside the viewport were not fully rendered when
the first frame entered the visible area. This caused incomplete or cut-off frames
during the initial stages of the animation. The issue was more noticeable on low-end
devices(RPi) or during slow-motion or frame-by-frame analysis of the recordings on x86.

With this change the visible rect is extended with the elements which are currently
not fully visible but they will be in a future bacause of animations.

* Source/WebCore/platform/graphics/texmap/coordinated/CoordinatedGraphicsLayer.cpp:
(WebCore::CoordinatedGraphicsLayer::transformedVisibleRectIncludingFuture):
(WebCore::CoordinatedGraphicsLayer::updateContentBuffers):
(WebCore::CoordinatedGraphicsLayer::computeTransformedVisibleRect):
* Source/WebCore/platform/graphics/texmap/coordinated/CoordinatedGraphicsLayer.h: