webgl context objects leak fix
Changing order of operations in ~WebGLRenderingContextBase
Otherwise, WebGLContextGroup will not be able to release some of the
'group shared' resources that still have attachements when being removed
via WebGLSharedObject destructor - it happens eg. for
WebGLVertexArrayObjectBase objects that have buffers set on
destruction time.

The current behaviour in ~WebGLRenderingContextBase:

- first, the reference to some WebGLSharedObject (like m_boundArrayBuffer)
   is nullified
- since this decreases refcnt to 0, destructor for this WebGLSharedObject is
  executed - this goes via WebGLObject::runDestructor() -> WebGLObject::deleteObject.
- WebGLObject::deleteObject checks for attachement count (m_attachmentCount)
  and if it's not 0 it will not execute deleteObjectImpl (where the actual
  removal from the context takes place).
- So the reference to the object is destructed, and ~WebGLSharedObject removes
  the object from m_contextGroup as well. At this point the resource is leaked.

With the fix, m_contextGroup->removeContext(*this) is executed at the beginning
of ~WebGLRenderingContextBase - this calls detach() on all the objects
(WebGLContextGroup::detachAndRemoveAllObjects) so is possible to release them.